code: be more explicit with RxPlayer's event types
Most modules internal to the RxPlayer emit some kind of events.

Due to its architecture, modules are sometimes contained in bigger
modules (the most impressive example would be the `StreamOrchestrator`,
which "orchestrates" multiple `PeriodStreams`, which switch between
`AdaptationStreams` which switch between `RepresentationStreams`).
Those outer-facing "super-modules" most often simply combine the events
of all sub-modules so they can be handled at a higher level as if they
were all coming from the same place (the `StreamOrchestrator` in our
example).

This way of doing things for a player has been shown to be relatively
maintainable and easy to update.

But it still in my opinion has some drawbacks:

  - first, the way those events were typed was hard to read. The types
    of those super-modules's events often redirected to the
    corresponding sub-modules type definitions.

    Because another module should not care about how that module is
    implemented, and most-of-all because multiple levels of indirection
    when browsing the code is a bad experience, I chose to impose flat
    type definitions when it comes to modules.

    This means that the return types of each module, which defines which
    events a module send, now directly link to the corresponding events
    with no level of indirection (other than the return type
    definition).

  - Second, which module had to react to which event could be very
    un-clear for a newcomer.

    For example a `"needs-media-source-reload"` was bubbled up until the
    API (the last possible step), even if it has always been already
    handled by the Init.

    I tried here to not bubble-up events which have no use being sent to
    outer modules.