source-buffer: remove inventory calculation logic from the QueuedSourceBuffer
-- Basic context about the modules involved

This modification concerns both the `Stream` module (the part of the
code deciding which segments should be loaded then pushed) and the
source-buffers (the part of the code handling browser-side media buffers
known as `SourceBuffers`).

The latter containing among other things both a `QueuedSourceBuffer`, an
abstraction over native and custom-defined SourceBuffers and a
`SegmentInventory` which will store information about which segments (as
in which quality, which track etc.) are actually present right now in
those.

-- What was the previous status

Just after pushing a segment, the `SegmentInventory` receives
information about it, so it can complete its internal representation
of what is contained in the linked SourceBuffer.

To do that, it relies notably on a `start` and `end` today refined two
times:

  1. First by the `transports` module, which will parse those
     information from the segments themselves and apply to it any
     possible offsets (such as the `timestampOffset`).

  2. Then by the `QueuedSourceBuffer`, which applies on it the
     append windows.

This behavior was bad for two reasons:

  1. On the `QueuedSourceBuffer` point of view, it makes no sense that
     it should apply both append windows on those time but not other
     SourceBuffer-related modifications such as the `timestampOffset`

     It was however necessary due to the fact that security paddings
     are added to those append windows to avoid removing too much data
     from segments on too-aggressive browsers.

     Setting those security paddings in the `transports` code make no
     sense in the whole RxPlayer architecture.

  2. In a future version of the RxPlayer where we might want to re-define
     one different type of `QueuedSourceBuffer` class for each type of
     buffers, that same logic have to be applied to all of them.

     This was actually the main reason behind that update, as
     low-latency contents may force us to update our source-buffers code
     for complex reasons.

-- What this commit does

This update moves the task of calculating the start and end of a given
segment from append windows for the `SegmentInventory`, from the
`QueuedSourceBuffer` to the `Stream`.

This also meant that security paddings added to append windows also had
to be moved there (which might still seems out of place, you're the judge
here).